# Add shallow cloning for Tokens to preserve the AST after render