--- a/lib/renderer.js
+++ b/lib/renderer.js
@@ -29,7 +29,10 @@ default_rules.code_block = function (tokens, idx /*, options, env */) {
 
 
 default_rules.fence = function (tokens, idx, options, env, slf) {
-  var token = tokens[idx],
+  // We will make modifications of the attributes object of this
+  // token while processing, so we obtain a shallow clone as to
+  // prevent modifying the token in the token list.
+  var token = tokens[idx].clone(),
       info = token.info ? unescapeAll(token.info).trim() : '',
       langName = '',
       highlighted;