@cbreeden 2016-06-19
It also should be noted that this is a breaking change, so it probably shouldn't be merged without considerable thought.