@puzrin 2016-06-19
Tokens stream must be immutable on render phase. Could you point me a line in source that breaks this principle?