@cbreeden 2016-06-19
It happens [in this call](https://github.com/markdown-it/markdown-it/blob/master/lib/renderer.js#L39).  When we [copy the token](https://github.com/markdown-it/markdown-it/blob/master/lib/renderer.js#L32) we essentially have only taken a pointer since objects are mutable, and then then when we modify the `attrs`, being a list, is also mutable and so all of these changes follow through to the token stream.  At least this is my understanding.  I verified this using a debugger.