@puzrin 2016-06-19
Got it. That's a bug that should be fixed. But i'm not sure it's needed to add `.clone()`. Created a separate issue to think about. Looks like `attrJoin()` is a design mistake.