@puzrin 2016-06-19
Found. `.attrJoin()` is ok, but is was a mistake to use it in renderer. Need to rewrite that function without `.attrJoin` and that will work without breaking changes.