@cbreeden 2016-06-19
I agree, using a `.clone()` is probably overkill.  I'll go ahead and close this, but I won't delete the branch if for now if we need to reference it for some reason later on.